pyproject.toml
[build-system]
requires = ["setuptools>=70.0.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pxbackup-ansible-runner"
version = "0.1.0"
description = "PX-Backup Ansible Runner"
requires-python = ">=3.11"
dependencies = [
    "Flask==3.0.2",
    "aiohttp>=3.9.4",
    "beaker>=20.1",
    "flask-caching>=2.3.1",
    "flask-cors>=4.0.1",
    "flask-limiter>=3.5.0",
    "gunicorn>=23.0.0",
    "cryptography>=42.0.5",
    "ecdsa>=0.19.1",
    "python-jose>=3.3.1",
    "okta-jwt-verifier==0.2.7",
    "python-keycloak==3.7.0",
    "PyJWT==2.8.0",
    "pydantic==2.5.2",
    "marshmallow==3.20.1",
    "prometheus-client==0.19.0",
    "structlog==23.2.0",
    "prometheus-flask-exporter==0.22.4",
    "redis==5.0.3",
    "psycopg2-binary==2.9.9",
    "SQLAlchemy==2.0.23",
    "alembic==1.13.0",
    "asyncpg==0.29.0",
    "hvac==2.1.0",
    "kubernetes>=29.0.0",
    "ansible-runner==2.3.4",
    "gitpython==3.1.42",
    "asyncio==3.4.3",
    "httpx==0.25.2",
    "grpcio==1.62.1",
    "python-dotenv==1.0.0",
    "tenacity==8.2.3",
    "psutil==5.9.6",
    "APScheduler==3.10.4",
    "PyYAML==6.0.1",
    "virtualenv>=20.26.6",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "--strict-markers"
markers = [
    "integration: marks tests as integration tests",
]

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
# A regex preceded with ^/ will apply only to files and directories
# in the root of the project.
^/tests/data/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 100
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip = ["tests/data"]
